treewide: refactor
Refactor mostly focusing on future expansion of
gui.
