+++ src/gui/performance_gui.rs
@@ -1,26 +1,17 @@
 use std::collections::VecDeque;
 
-use bevy::{input::mouse::MouseWheel, render::render_resource::encase::private::Length};
+use bevy::prelude::*;
 use bevy_egui::{
     egui::{self, Slider},
-    EguiContexts, EguiPlugin,
+    EguiContexts,
 };
 use egui_plot::{Line, Plot, PlotPoints};
 
-use crate::*;
+pub struct PreformanceGuiPlugin;
 
-pub struct GuiPlugin;
-
-impl Plugin for GuiPlugin {
+impl Plugin for PreformanceGuiPlugin {
     fn build(&self, app: &mut App) {
-        app.add_plugins(EguiPlugin)
-            .add_systems(Update, performance_gui)
-            .add_systems(
-                PreUpdate,
-                absorb_egui_inputs
-                    .after(bevy_egui::input::write_egui_input_system)
-                    .before(bevy_egui::begin_pass_system),
-            )
+        app.add_systems(Update, performance_gui)
             .insert_resource(PerformanceData {
                 frame_time: [0.0].into(),
             })
@@ -50,7 +41,7 @@ fn performance_gui(
         .frame_time
         .push_back(1000.0 * time.delta_secs());
 
-    while performance_data.frame_time.length() > gui_settings.history_to_show as usize {
+    while performance_data.frame_time.len() > gui_settings.history_to_show as usize {
         performance_data.frame_time.pop_front();
     }
 
@@ -84,38 +75,4 @@ fn performance_gui(
 
         ui.add(Slider::new(&mut gui_settings.history_to_show, 0..=500).text("history to show"))
     });
-}
-
-// source:
-// https://github.com/vladbat00/bevy_egui/issues/47#issuecomment-2368811068
-fn absorb_egui_inputs(
-    mut contexts: bevy_egui::EguiContexts,
-    mut mouse: ResMut<ButtonInput<MouseButton>>,
-    mut mouse_wheel: ResMut<Events<MouseWheel>>,
-    mut keyboard: ResMut<ButtonInput<KeyCode>>,
-) {
-    let ctx = contexts.ctx_mut();
-    if !(ctx.wants_pointer_input() || ctx.is_pointer_over_area()) {
-        return;
-    }
-    let modifiers = [
-        KeyCode::SuperLeft,
-        KeyCode::SuperRight,
-        KeyCode::ControlLeft,
-        KeyCode::ControlRight,
-        KeyCode::AltLeft,
-        KeyCode::AltRight,
-        KeyCode::ShiftLeft,
-        KeyCode::ShiftRight,
-    ];
-
-    let pressed = modifiers.map(|key| keyboard.pressed(key).then_some(key));
-
-    mouse.reset_all();
-    mouse_wheel.clear();
-    keyboard.reset_all();
-
-    for key in pressed.into_iter().flatten() {
-        keyboard.press(key);
-    }
 }
